xml/parsing/manakai/cdata-2.dat

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!ELEMENT x ANY>
]><x><![CDATA[ ]]></x>
#errors
#document escaped
| <!DOCTYPE x>
|   <!ELEMENT x ANY>
| <{} x>
|   " "

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!ELEMENT x ANY>
]><x><![CDATA[]]></x>
#errors
#document escaped
| <!DOCTYPE x>
|   <!ELEMENT x ANY>
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!ELEMENT x ( #PCDATA | y)*>
]><x><![CDATA[ ]]></x>
#errors
#document escaped
| <!DOCTYPE x>
|   <!ELEMENT x (#PCDATA | y)*>
| <{} x>
|   " "

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!ELEMENT x ( #PCDATA | y)*>
]><x><![CDATA[]]></x>
#errors
#document escaped
| <!DOCTYPE x>
|   <!ELEMENT x (#PCDATA | y)*>
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!ELEMENT x ( PCDATA | y)*>
]><x><![CDATA[ ]]></x>
#errors
3;6;xml:CDATA section not allowed by cm;;;m
#document escaped
| <!DOCTYPE x>
|   <!ELEMENT x (PCDATA | y)*>
| <{} x>
|   " "

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!ELEMENT x ( PCDATA | y)*>
]><x><![CDATA[]]></x>
#errors
3;6;xml:CDATA section not allowed by cm;;;m
#document escaped
| <!DOCTYPE x>
|   <!ELEMENT x (PCDATA | y)*>
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!ELEMENT x EMPTY>
]><x><![CDATA[ ]]></x>
#errors
3;3;xml:start tag:EMPTY;x;;s
3;6;xml:CDATA section not allowed by cm;;;m
#document escaped
| <!DOCTYPE x>
|   <!ELEMENT x EMPTY>
| <{} x>
|   " "

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!ELEMENT x EMPTY>
]><x><![CDATA[]]></x>
#errors
3;3;xml:start tag:EMPTY;x;;s
3;6;xml:CDATA section not allowed by cm;;;m
#document escaped
| <!DOCTYPE x>
|   <!ELEMENT x EMPTY>
| <{} x>
